feat: add support for PUID and PGID environment variables
This commit removes the hardcoded MUMBLE_UID and MUMBLE_GID and instead
adopts the commonly used PUID and PGID environment variables for setting
the desired UID:GID for mumble-server to run as. For backwards-compatibility
they still default to 10000:10000.

Also, a new variable `MUMBLE_CHOWN_DATA` is used to define whether to chown
`/data` and its contents or not.

Fixes #34

Signed-off-by: Nita Vesa <nita.vesa@outlook.com>